fix: namespaced time adjustment(s)
