+++ app/graphql/types/tfl/entities/journey_planner/search_criteria_type.rb
@@ -5,7 +5,7 @@ module Types
         class SearchCriteriaType < Types::BaseObject
           field :dateTime, String, null: true
           field :dateTimeType, Types::DateTimeType, null: true
-          field :timeAdjustments, Types::TimeAdjustmentsType, null: true
+          field :timeAdjustments, Types::Tfl::Entities::JourneyPlanner::TimeAdjustmentsType, null: true
         end
       end
     end
