--- app/graphql/types/time_adjustment_type.rb
@@ -1,8 +1,0 @@
-module Types
-  class TimeAdjustmentType < Types::BaseObject
-    field :date, String, null: true
-    field :time, String, null: true
-    field :timeIs, String, null: true
-    field :uri, String, null: true
-  end
-end
