--- app/graphql/types/time_adjustments_type.rb
@@ -1,8 +1,0 @@
-module Types
-  class TimeAdjustmentsType < Types::BaseObject
-    field :earlier, Types::TimeAdjustmentType, null: true
-    field :earliest, Types::TimeAdjustmentType, null: true
-    field :later, Types::TimeAdjustmentType, null: true
-    field :latest, Types::TimeAdjustmentType, null: true
-  end
-end
